==> Prog/impact-estimates/04_vacant-lots-fairfax.Rmd <==
---
title: "Regional Housing Framework"
subtitle: "Vacant lots analysis for Fairfax County"
author: ""
output:
  html_document:
    number_sections: FALSE
    self_contained: TRUE
    code_folding: show
    toc: TRUE
    toc_float: TRUE
    css: ../../www/web_report.css
    editor_options:
      chunk_output_type: console
---

<link rel="stylesheet" href="//fonts.googleapis.com/css?family=Lato" />

Library: RegHsg

Project: Regional Housing Framework

Author: Chris Davis, adapting from Sarah Strochak

Version: R 3.5.1, RStudio 1.1.423

Last updated `r format(Sys.time(), '%B %d, %Y')`

Environment: Local Windows session (desktop)


```{r rmarkdown-setup, echo = FALSE}
knitr::opts_chunk$set(warning = FALSE)
knitr::opts_chunk$set(message = FALSE)
# knitr::opts_chunk$set(eval = FALSE)
```

## Description

The goal of this analysis is to determine how many units could be added by filling vacant lots with the most-dense developments allowed under the current zoning codes.
The two steps are:

1. Merge cleaned vacant lots data with zoning code data.
2. Calculate how many units are allowed one each vacant lot.

## Set-up
Load libraries and functions
```{r setup}
library(tidyverse)
library(urbnthemes)
library(sf)
library(magrittr)
library(tigris)
set_urbn_defaults("print")

source("../../Macros/read-fairfax.R")

```

Create directory for data exports
```{r dir}
if (!dir.exists("../../Data")) {
  dir.create("../../Data")
}

if (!dir.exists("L:/Libraries/RegHsg/Prog/Tables")) {
  dir.create("L:/Libraries/RegHsg/Prog/Tables")
}
```

Set FIPS code and filepath name
```{r fips}

currentfips <- "51059"
filepath <- "fairfax"


```

Load in cleaned Black Knight data for `r str_to_title(filepath)` County.

```{r read}
if (!exists("jur")) {
  jur <- read_fairfax()
} else {
  warning(filepath, " data already read in")
}

```


## Clean zoning code information

### Move file 

First- **manually** move file from Box to the `zoning-codes` directory in the `Doc` folder on the L drive. Name file accordingly:
"jursidiction-zoning-code.csv"

### Read file

Read in zoning code, rename variables so that they are machine readable

```{r file-check}

if (!file.exists(paste0("L:/Libraries/RegHsg/Doc/zoning-codes/",
                              filepath,
                              "-zoning-code.csv"))) {
  
  stop("Move zoning file for jurisdiction to L:/Libraries/RegHsg/Doc/zoning-codes directory")
}

```

```{r read-zoning-csv}
z <- cols(
  code = col_character(),
  name = col_character(),
  dwelling_type = col_character(),
  ave_lot_area = col_integer(),
  lot_area = col_character(),
  max_dens = col_integer(),
  height = col_integer(),
  density = col_integer()
)

zoningcode <- read_csv(paste0("L:/Libraries/RegHsg/Doc/zoning-codes/",
                              filepath,
                              "-zoning-code.csv"),
                       col_types = z)

rm(z)

```

### Creating dwelling_type hierarchy

For the vacant lots analysis, we want to determine the `dwelling_type` that allows for the most dense construction for that zone. 

First, we look at all the options, and filter out all the observations marked as "other", since we cannot be sure what property type that refers to. We will also filter out categories that are dependent on adjacent lot lines.

```{r tab-dwellings}

zoningcode %>% 
  count(dwelling_type) %>% 
  knitr::kable()

```

Count the number of unique zoning codes to make sure we won't get rid of any codes entirely.

```{r count-zoning}
length(unique(zoningcode$code))

```


Since the variables are not always comparable, the following order, from most to least dense, applies:

                          
                          
1. Multiple family dwellings (ADU)
2. Single family attached (ADU)
3. Single family attached
4. Single family detached (ADU)
5. Single family detached
6. Dwellings (all structures -- standard)
7. Mobile home park


### Create factor variable
To use this hierarchy, we transform the `dwelling_type` variable into a factor, assign the factors to the appropriate level, (1-6, as listed above).
We will also create a designation for dwelling types that are restricted to a certain amount of units (one unit for "One Family Dwelling" and one for two units for "Duplex"). 

**NOTE** This will be VERY jurisdiction specific.

**NOTE** If there are no one or two unit limitations, still make the variable and set them all equal to zero. This allows the functionality of the calculation to be more uniform across counties.

```{r dwelling-factor}

zoningcode <- zoningcode %>% 
  mutate(dwelling_type = factor(dwelling_type,
                                levels = c(# "single family detached (ADU)",
                                           "multiple family dwellings (ADU)",
                                           "single family attached (ADU)", 
                                           "single family attached",
                                           "single family detached (ADU)",
                                           "single family detached",
                                           "dwellings (all structures -- standard)",
                                           "mobile home park"))) %>% 
  mutate(one_unit_flag = ifelse(dwelling_type == "single family detached (ADU)" |
                                  dwelling_type == "single family detached",
                                1,
                                0)) #  %>% 
  # mutate(two_unit_flag = ifelse(dwelling_type == "single family attached (ADU)" |
  #                                 dwelling_type == "single family attached",
  #                               1,
  #                               0))

# count(zoningcode, dwelling_type)
# count(zoningcodex, dwelling_type)
# count(zoningcodex, one_unit_flag)

```


## Clean vacant lots data

See how many lots have missing key variables

```{r missing}

jur %>% 
  mutate(num = NA) %>% 
  group_by(category_detail) %>% 
  summarize_at(vars(propaddress, lotsize_sf, zoning, num),
               ~ sum(is.na(.))) %>% 
  mutate_at(vars(propaddress, lotsize_sf, zoning),
            .funs = funs(percent = scales::percent(. / num))) %>% 
  knitr::kable(caption = "Number and percent of missing values")

```


### Assessment values

The difference between the land value, improvement value, and total value can help us determine if lots are really vacant.

```{r check-assess}

jur %>% 
  group_by(category_detail) %>% 
  summarize(nprops = n(),
            imp_zero = sum(assessedimprovementvalue == 0, na.rm = TRUE),
            imp_na = sum(is.na(assessedimprovementvalue))) %>% 
  mutate(imp_none = imp_zero + imp_na,
         percent_none = imp_none / nprops) %>% 
  knitr::kable()
  

```


Most of the vacant properties have either a 0 or a missing value for the improvement value, which is what we would expect. We can take a closer look at properties that have a non-zero value and decide if we need to re-categorize them as not really vacant- if there is a structure there, we do not want to include it in this calculation.

```{r not-so-vacant}

jur %>% 
  filter(category=="vacant",
         assessedimprovementvalue > 0, 
         !is.na(assessedimprovementvalue)) %>% 
  mutate(p_total = assessedimprovementvalue / totalassessedvalue) %>% 
  group_by(category_detail) %>% 
  summarize(n = n(),
            mean = mean(assessedimprovementvalue),
            median = median(assessedimprovementvalue),
            min = min(assessedimprovementvalue),
            max = max(assessedimprovementvalue),
            p_total = median(p_total)) %>% 
  knitr::kable()

```

Look at the distribution of improved value as a percent of total value.

```{r scatter-not-vacant, echo=FALSE}
jur %>% 
  filter(category=="vacant",
    assessedimprovementvalue > 0, 
         !is.na(assessedimprovementvalue)) %>% 
  mutate(p_total = assessedimprovementvalue / totalassessedvalue) %>% 
  ggplot() +
  geom_histogram(mapping = aes(p_total),
                 bins = 30) +
  scale_x_continuous(labels = scales::percent,
                     breaks = c(0:10 *.1)) +
  scale_y_continuous(expand = c(0,0)) +
  labs(x = "Percent of total assessed value that is derived from improvement value",
       y = NULL)
```

We can recategorize these observations as "vacant with structure", and will not include these in the calculation. 

```{r not-vacant-recat}

jur_2 <- jur   %>% 
  mutate(category = ifelse(assessedimprovementvalue > 0 & 
                           !is.na(assessedimprovementvalue) &
                             category == "vacant",
                           "vacant with structure",
                           category))
count(jur_2, category)


```

Need to clean the zoning variable to get rid of parenthesis in BlackKnight data and make readable. 

```{r}
jur_2 %<>% 
  mutate(newZoning = case_when(
    str_detect(zoning, "\\(") ~ substring(zoning, 1, regexpr("\\(", zoning) - 1),
    TRUE ~ zoning)) %>%
  mutate(newZoning = case_when(
    newZoning=="PRC" ~ "PRC",
    newZoning=="PDC" ~ "PDC",
    newZoning=="PTC" ~ "PTC",
    newZoning=="RC" ~ "RC",
    str_ends(newZoning, "C") ~ str_sub(newZoning, 1, str_length(newZoning)-1),
    TRUE ~ newZoning)) 
```


We make a flag for missing coordinates.

```{r}

jur_3 <- jur_2 %>% 
  mutate(missing_coord = ifelse(is.na(lat) | is.na(long),
                                1,
                                0))
```

### Missing zoning variables

To determine zoning codes for addresses that have a missing zoning designation or multiple zoning designations, we will use a spatial join to the county's zoning layer.

Read in zoning layer, set CRS

```{r read-zoning}
zdir <- paste0("L:/Libraries/RegHsg/Maps/",
               filepath,
               "-zoning-map")
zfile <- str_sub(list.files(zdir), end = -5) %>% unique()

zon <- read_sf(dsn = zdir,
               layer = zfile)

st_geometry(zon)
zon <- st_transform(zon, crs = 4326)

```




  
Convert vacant lots to SF, set CRS. To convert to SF, we need to use only the observations with non-missing longitude and latitude. We separate these off and recombine after the spatial join.



```{r convert-sf}

jursf <- jur_3 %>% 
  filter(missing_coord==0) %>% 
  st_as_sf(coords = c("long", "lat")) %>% 
  st_set_crs(st_crs(zon))

```



Filter out Herndon, Vienna, and Clifton as these jurisdictions do not fall under Fairfax County zoning.

```{r}

options(tigris_use_cache = TRUE)

places <- places("Virginia") %>% 
  st_as_sf() %>% 
  st_transform(st_crs(zon)) %>% 
  filter(NAME %in% c("Herndon", "Vienna", "Clifton")) %>% 
  select()


jursf_temp <- jursf %>% 
  mutate(notIsPlace= sapply(st_intersects(jursf, places),
                            function(x){length(x)==0}))

# check that these jurisdictions are in Clifton, Herndon, or Vienna
jursf_temp %>% 
  filter(notIsPlace == FALSE) %>% 
  count(propcity)

# count many rows were removed
removed <- nrow(jursf) - nrow(filter(jursf_temp, notIsPlace == TRUE)) 
removed

# remove those jurisdictions
jursf_temp1 <- jursf_temp %>% 
  filter(notIsPlace == TRUE) %>% 
  mutate(id = seq(1, nrow(.), 1))

```



Spatial join to zoning layer

```{r spatial-join}

jurjoin <- st_join(jursf_temp1, zon, join = st_intersects)

```

23 observations are gained through this join - these are added because points are falling in more than one zone.


After the join, we convert back to a dataframe, for ease of computations.
```{r back-to-df}

jur_4<- st_set_geometry(jurjoin, NULL)

```

Count how many observations did not join. This will throw an error if more than 5% did not match- in this case, check the underlying shapefiles for compatability.

```{r count-spatial}

paste0(sum(is.na(jur_4$ZONECODE)), " observations didn't join- ",
      round(sum(is.na(jur_4$ZONECODE)) / nrow(jur_4) * 100, 2), "%")

if (sum(is.na(jur_4$ZONECODE)) / nrow(jur_4) > .05) {
  stop("More than 5% of observations could not be spatially joined.")
}
      
```


Drop new observations that were added in the join

```{r}
tempjur <- jur_4 %>% 
  mutate(newId = seq(1, nrow(jur_4), 1))

tempjur1 <- tempjur %>% 
  group_by(id) %>% 
  count() %>% 
  arrange(desc(n)) %>% 
  filter(n>1)

tempjur2 <- left_join(tempjur1, tempjur %>% select(id, newId),
                       by = "id")

tempjur3 <- tempjur2[seq(1, nrow(tempjur2), 2), ]

jur_5 <- tempjur %>% filter(!newId%in%tempjur3$newId) %>% select(-id, -newId)

```


Test how many have different results from the Black Knight zoning variable.

```{r test-spatial}

jur_5 %>% 
  group_by(is.na(newZoning)) %>% 
  count(newZoning == ZONECODE)

```

Doesnt look like the spatial join help in creating new variable. 

Take a closer look how many and what percent are different "just because"- not because they didn't merge, have missing values, or have duplicate values.

Note: For fairfax, "/" exists in many codes that do not contain duplicates. The only duplicates that are identified are for zones that have multiple "\""'s. 

```{r test-spatial-again}

jur_5 %>% 
  filter(!is.na(newZoning), 
         !is.na(ZONECODE)) %>% 
  mutate(different_flag = ifelse(newZoning != ZONECODE,
                                 1,
                                 0)) %>% 
  summarize(different = sum(different_flag, na.rm = TRUE)) %>% 
  mutate(percent_different = paste0(round(different / nrow(jur_3) *100, 2), "%"))





```

Many are different for some jurisdictions because the Black Knight zoning has two zoning codes, separated by a slash. We can test how many of these instances have the spatially joined zoning designation in the Black Knight zoning variable. However, this is not true for Fairfax. The slashes denote jurisdictions that either have "w/" to mean "with" or have some other use. 

```{r}
sames <- jur_5 %>% filter(newZoning == ZONECODE)
differents <- jur_5 %>% filter(newZoning != ZONECODE)
others <- jur_5 %>% filter(is.na(newZoning)|is.na(ZONECODE))


differents %<>% mutate(ZONECODE = gsub(pattern="-", replacement = "", ZONECODE))

sames1 <- differents %>% 
  filter(newZoning == ZONECODE)
differents1 <- differents %>% 
  filter(newZoning != ZONECODE)


differents1 %<>% mutate(ZONECODE = case_when(
                     ZONECODE == "R3" ~ "R-3",
                     ZONECODE == "R1" ~ "R-1",
                     ZONECODE == "C8" ~ "C-8",
                     ZONECODE == "R2" ~ "R-2",
                     ZONECODE == "C3" ~ "C-3",
                     ZONECODE == "R8" ~ "R-8",
                     ZONECODE == "I4" ~ "I-4",
                     ZONECODE == "I5" ~ "I-5",
                     ZONECODE == "I3" ~ "I-3",
                     ZONECODE == "PDH2" ~ "PDH-2",
                     ZONECODE == "PDH5" ~ "PDH-5",
                     ZONECODE == "C5" ~ "C-5",
                     ZONECODE == "R20" ~ "R-20",
                     ZONECODE == "PDH12" ~ "PDH-12",
                     ZONECODE == "C7"~"C-7",
                     ZONECODE == "C4" ~ "C-4",
                     ZONECODE == "I1" ~ "I-1",
                     ZONECODE == "C6" ~ "C-6",
                     ZONECODE == "R4" ~ "R-4",
                     ZONECODE == "C2" ~ "C-2",
                     ZONECODE == "R5" ~ "R-5",
                     ZONECODE == "R12" ~ "R-12",
                     ZONECODE == "PDH4" ~ "PDH-4",
                     ZONECODE == "PDH3" ~ "PDH-3",
                     ZONECODE == "PDH20"~ "PDH-20",
                     ZONECODE == "PDH30" ~ "PDH-30",
                     ZONECODE == "PDH8" ~ "PDH-8",
                     ZONECODE == "PDH40" ~ "PDH-40",
                     ZONECODE == "I6" ~ "I-6",
                     ZONECODE == "I2" ~ "I-2",
                     ZONECODE == "PDH1" ~ "PDH-1",
                     ZONECODE == "PDH16" ~ "PDH-16",
                     ZONECODE == "R30" ~ "R-30",
                     ZONECODE == "R16" ~ "R-16",
                     TRUE ~ ZONECODE
)) %>% mutate(newZoning=case_when(
  newZoning == "PLANNED RES MIXED USE"     ~ ZONECODE,
  newZoning == "R-10 SINGLE FAM 10000 3+D" ~ ZONECODE,
  newZoning == "R12"                       ~ "R-12",
  newZoning == "R"                         ~ "RC",
  TRUE                                     ~ newZoning))


others %<>%  mutate(ZONECODE = case_when(
                     ZONECODE=="R3" ~ "R-3",
                     ZONECODE=="R1" ~ "R-1",
                     ZONECODE=="C8" ~ "C-8",
                     ZONECODE=="R2" ~ "R-2",
                     ZONECODE=="C3" ~ "C-3",
                     ZONECODE=="R8" ~ "R-8",
                     ZONECODE=="I4" ~ "I-4",
                     ZONECODE=="I5" ~ "I-5",
                     ZONECODE=="I3" ~ "I-3",
                     ZONECODE=="PDH2" ~ "PDH-2",
                     ZONECODE=="PDH5" ~ "PDH-5",
                     ZONECODE=="C5" ~ "C-5",
                     ZONECODE=="R20" ~ "R-20",
                     ZONECODE=="PDH12" ~ "PDH-12",
                     ZONECODE=="C7"~"C-7",
                     ZONECODE=="C4" ~ "C-4",
                     ZONECODE=="I1" ~ "I-1",
                     ZONECODE=="C6" ~ "C-6",
                     ZONECODE=="R4" ~ "R-4",
                     ZONECODE=="C2" ~ "C-2",
                     ZONECODE=="R5" ~ "R-5",
                     ZONECODE=="R12" ~ "R-12",
                     ZONECODE=="PDH4" ~ "PDH-4",
                     ZONECODE=="PDH3" ~ "PDH-3",
                     ZONECODE=="PDH20"~ "PDH-20",
                     ZONECODE=="PDH30" ~ "PDH-30",
                     ZONECODE=="PDH8" ~ "PDH-8",
                     ZONECODE=="PDH40" ~ "PDH-40",
                     ZONECODE=="I6" ~ "I-6",
                     ZONECODE=="I2" ~ "I-2",
                     ZONECODE=="PDH1" ~ "PDH-1",
                     ZONECODE=="PDH16" ~ "PDH-16",
                     ZONECODE=="R30" ~ "R-30",
                     ZONECODE=="R16" ~ "R-16",
                     TRUE ~ ZONECODE
)) %>%  mutate(newZoning=case_when(
  newZoning == "PLANNED RES MIXED USE"      ~ ZONECODE,
  newZoning =="R-10 SINGLE FAM 10000 3+D"   ~ ZONECODE,
  newZoning =="R12"                         ~ "R-12",
  newZoning =="R"                           ~ "R-1",
  TRUE                                      ~ newZoning))

differents2 <- differents1 %>% 
  filter(newZoning != ZONECODE)





jur_5a <- differents1 %>% 
         bind_rows(sames) %>%
         bind_rows(sames1) %>%
         bind_rows(others)


```

Fix newZoning that has weird residential values

```{r}
residential_weird <- c("RES ",
                       "RD",
                       "MULTI-FAM W/MIX COM/INDUS",
                       "R-10",
                       "RTC-10",
                       "MIX AG AND CONSRV W/RES/A",
                       "RP",
                       "RS-16",
                       "R-15",
                       "RS-10",
                       "MULTI-FAM W/MIX OTHER RES",
                       "RM-1",
                       "RS-12.5",
                       "MULTI FAM W/MIX MULTI FAM",
                       "R10",
                       "RES W OTHER RES ZONING",
                       "RM-2")

jur_6 <- jur_5a %>% mutate(newZoning = case_when(
  newZoning %in% residential_weird ~ ZONECODE,
  TRUE ~ newZoning
))


```


```{r slashes2}

jur_6 %>% 
  filter(!is.na(newZoning), 
         !is.na(ZONECODE)) %>% 
  mutate(different_flag = ifelse(newZoning != ZONECODE,
                                 1,
                                 0)) %>% 
  summarize(different = sum(different_flag, na.rm = TRUE)) %>% 
  mutate(percent_different = paste0(round(different / nrow(jur_4) *100, 2), "%"))
```

Recombine with observations with missing coordinates, check that number of rows is the same.

```{r get-final}

jurfinal <- jur_3 %>% 
  filter(is.na(lat) | is.na(long)) %>% 
  bind_rows(jur_6)

stopifnot((nrow(jurfinal)+removed) == nrow(jur_3))

```

### Dual zoning codes

Fairfax does not have dual zoning codes.



## Merge with cleaned data

Now that we have selected the correct records from the zoning code file, we can merge with the cleaned property records data. First, we have to do a bit of cleaning of these codes, including getting rid of instances where there is a tie, and correcting some of the notation. We will create a new variable in order to archive the original `zoning` variable.



```{r bk-clean}

jurfinal <- jurfinal %>% 
  mutate(zoning_code = case_when(
    newZoning == "RA" ~ "R-A",
    newZoning == "RC" ~ "R-C",
    newZoning == "RE" ~ "R-E",
    newZoning == "R-15 SINGLE FAM 15000 2+D" ~ "R-15",
    newZoning == "R-10 SINGLE FAM 10000 3+D" ~ "R-10",
    newZoning == "R12" ~ "R-12",
    newZoning == "R10" ~ "R-10", 
    TRUE ~ newZoning))
           
```

```{r}
(jurfinal$zoning_code%>%unique)[!(jurfinal$zoning_code %>% unique %in% (zoningcode$code %>%unique))] %>% sort 
```

```{r}
(zoningcode$code%>%unique)[!(zoningcode$code %>% unique %in% (jurfinal$zoning_code%>%unique))] 
```



Merge on zoning code variable. Test that all variables merged.

```{r zoning-merge}

zoningdata <- left_join(jurfinal, zoningcode, by = c("zoning_code"="code"))

```



There are `r ` observations that did not have a match in the zoning code. Below, we see which zoning codes were in the Black Knight data but NOT in the zoning code data.

```{r antijoin}

anti_join(jurfinal, zoningcode, by = c("zoning_code"="code")) %>% 
  count(zoning_code) %>% 
  arrange(desc(n)) %>%
  knitr::kable()

```


Categorize those lots that are not joining
```{r unjoined-categories}

planned<-c("PRC",
           "PDH-3",
           "PDH-4",
           "PDH-8",
           "PDH-2",
           "PDH-12",
           "PDC",
           "PDH-16",
           "PDH-5",
           "PDH-20",
           "PRM",
           "PDH-30",
           "PDH-1",
           "PDH-40",
           "PDH",
           "PLANNED RES MIXED USE",
           "PRM",
           "PTC")

notResidential <- c("I-5",
                    "C-2",
                    "C-8",
                    "C-6",
                    "C-3",
                    "I-4",
                    "C-5",
                    "I-3",
                    "COMM ",
                    "I-6",
                    "INDUSTRIAL ",
                    "C-7",
                    "INDUSTR ",
                    "PUBLIC SERVICE CORPORATIO",
                    "C-1",
                    "C-4",
                    "CD",
                    "I-I",
                    "I-2",
                    "CO",
                    "CS",
                    "COM ",
                    "ID",
                    "CM",
                    "CMP",
                    "C-1A",
                    "CCD",
                    "I-1",
                    "INDUST W/OTHER INDUST ZON"
)
  


tempTable <- anti_join(jurfinal, zoningcode, by = c("zoning_code"="code")) %>% 
  count(zoning_code) %>% arrange(desc(n)) %>% 
  mutate(type = case_when(
    zoning_code %in% planned ~ "planned",
    zoning_code %in% notResidential ~ "not residential",
    zoning_code %in% residential_weird ~ "residential weird",
    TRUE ~ "other"))

temperTable <- tempTable %>% 
  group_by(type) %>% 
  summarise(totalLots = sum(n, na.rm=T))

temperTable%>%
  knitr::kable()

```

Of those that do not match, `r paste0(round(temperTable %>% filter(type=="planned") %>% pull(totalLots) / temperTable$totalLots %>% sum, 2)*100, "%")` (`r temperTable %>% filter(type=="planned") %>% pull(totalLots)`) are in planning development districts, some of which have residential uses. However, developers must pass the development plan process before building in these districts. `r temperTable %>% filter(type == "residential weird") %>% pull(totalLots)` of those classified as "residential weird" are vacant.

```{r results = F}

planned_vacants<- jurfinal %>% 
  filter(zoning_code %in% planned,
         category=="vacant") %>%
  nrow()

```

There are `r planned_vacants` vacant lots in the aforementioned planned jurisdictions


Alter lotsize for a lot - at "KATELYN ZINN PL" - that has a lotsize that is far too high - size was in acres when it should have been feet.


```{r}

zoningdata_sub <- zoningdata %>% 
  mutate(lotsize_sf = ifelse(grepl("KATELYN ZINN PL", propaddress),
                                                            lotsize_sf/43560, 
                                                            lotsize_sf))

```


Filter out lots that are too small to build on and lots with missing zoning information, and select most dense zoning code for those that are big enough.

```{r filter-small}


zoningdata1 <- zoningdata_sub %>% 
  filter(!is.na(newZoning)) %>% 
  mutate(smallsize_flag = ifelse(lotsize_sf >= density, 0, 1)) %>% 
  mutate(smallsize_flag = ifelse(is.na(smallsize_flag), 0, smallsize_flag)) %>% 
  group_by(propaddress, smallsize_flag) %>% 
  slice(which.min(dwelling_type)) %>% 
  ungroup() %>% 
  group_by(propaddress) %>% 
  slice(which.min(smallsize_flag)) %>% 
  ungroup() %>%
  mutate(lotsize_sf = as.numeric(lotsize_sf),
         lot_area_min = gsub(lot_area, pattern=",", replacement = "") %>% str_trim() %>% as.numeric) 

# check observations
nrow(zoningdata1) + nrow(filter(zoningdata_sub, is.na(zoning_code)))

# this is the number of all parcels, including those with 
zoningdata_sub %>% group_by(propaddress) %>% count() %>% nrow()

# 
# count(zoningdata1, smallsize_flag)
# 
# t <- zoningdata1 %>% 
#   filter(smallsize_flag == 1)
# 
# count(zoningdata1, one_unit_flag)

```

## Calculate missing units

There are several variables we will use to calculate how many units we can put on each vacant lot.

1. Buildable area: this is `lotsize_sf` x the coverage ratio (`lot_coverage_max`)
2. Living area: this is the buildable area x the construction efficiency. The lower bound estimate is 70%, the upper bound estimate is 85%.
3. Units per floor: this is the living area divided by the average unit size. We use 940 sq. feet, based on the American Housing Survey 2017 data for the DC metro area.
4. Units allowed based on height limits: This is the units per floor x the number of stories.
3. Units allowed based on lot size: this is the lot area divided by `lotarea_per_unit`

For Fairfax, `lotarea_per_unit` is assumed to be density per acre, as this seems to be the main restriction in this code. 

We can create a function that will allow us to easily adjust these parameters and compare results.

```{r get-est}

#' Get maximum unit variables
#' @description this function generates the variables necessary to calculate the maximum number of units that can occupy one lot, per the by-right zoning code.
#'
#' @param dataset dataset with lot size and zoning variables 
#' @param construction_efficiency Assumption for construction efficiency
#' @param average_unit_size Assuption for average unit size- should be derived from the American Housing Survey. Given in square feet.
#'
#' @return returns a dataframe with new variables: `parameters`, `buildable_area`, `living_area`, `units_per_floor`, `floors`, `units_height`, `units_lotsize`, and `units_total`
#'
#' @examples
get_unit_variables <- function(dataset, 
                               construction_efficiency, 
                               average_unit_size) {
  
  dataset %>% 
    mutate(parameters = paste0("CE: ", construction_efficiency * 100,
                               "% , Unit size: ", average_unit_size,
                               " sf"),
           lot_coverage = .75,
           buildable_area = lotsize_sf * lot_coverage,
           living_area = buildable_area * construction_efficiency,
           units_per_floor = floor(living_area / average_unit_size),
           floors = floor(height / 10),
           units_height = units_per_floor * floors,
           units_density = floor(lotsize_sf / density),
           subset = ifelse(is.na(lot_area_min), NA, floor(lotsize_sf/density))) %>%   
      mutate_at(vars(units_height, units_density),
              funs(case_when(
                lotsize_sf < density ~ 0,
                one_unit_flag == 1 & is.na(.) ~ 1,
                one_unit_flag == 1 & subset > 1 & !is.na(subset) ~ subset,
                one_unit_flag == 1 & (subset < 1 | is.na(subset)) ~ 1,
                # two_unit_flag == 1 & . >= 2 & !is.na(subset) ~ 2 * subset,
                # two_unit_flag == 1 & . >= 2 & (subset < 1 | is.na(subset)) ~ 2,
                TRUE ~ .))) %>% 
      mutate(units_total = case_when(
             # lotsize_sf < lot_area_min ~ 0,
              lotsize_sf < density ~ 0,
              TRUE ~ pmin(units_density, units_height, na.rm = TRUE)))
    
}

#' Get maximum unit summary
#'
#' @param dataset dataset with lot size and zoning variables 
#' @param construction_efficiency Assumption for construction efficiency
#' @param average_unit_size Assuption for average unit size- should be derived from the American Housing Survey. Given in square feet.
#' @param dwelling_type_breakdown Options are TRUE or FALSE- indicates whether or not estimates should be broken down by `dwelling_type`
#'
#' @return Returns a dataframe with summary statistics on the maximum units that can be build on lots, per the by-right zoning code.
#'
#' @examples
get_unit_summary <- function(dataset, 
                             construction_efficiency, 
                             average_unit_size,
                             dwelling_type_breakdown) {

  x <- dataset %>% 
    get_unit_variables(construction_efficiency,
                       average_unit_size) %>%
    group_by(parameters) %>% 
    summarize(lots = n(),
              units_height = sum(units_height),
              units_density = sum(units_density),
              units_total = sum(units_total),
              units_per_lot = floor(sum(units_total / lots)))
             # units_area = sum(units_area),
             # units_acre = sum(units_acre)))
  
  y <- dataset %>% 
    get_unit_variables(construction_efficiency,
                       average_unit_size) %>%
    group_by(parameters, dwelling_type) %>% 
    summarize(lots = n(),
              units_height = sum(units_height),
              units_density = sum(units_density),
              units_total = sum(units_total),
              units_per_lot = floor(sum(units_total / lots))) %>%
          
    
    mutate(dwelling_type = as.character(dwelling_type))
  
  
  if (dwelling_type_breakdown == TRUE) {
   
      bind_rows(y, mutate(x, dwelling_type = "All"))  
  
  } else if (dwelling_type_breakdown == FALSE) {
      x
    }
}



```
    

### All lots

Run the function on the entire dataset, and export for soft sites analysis.

```{r units-all}

units <- get_unit_variables(dataset = zoningdata1,
                            construction_efficiency = .7,
                            average_unit_size = 940)

```


Write out for soft sites.
```{r write-units}

saveRDS(units,
        "L:/Libraries/RegHsg/Data/fairfax/fairfax-max-units.Rdata")

```


### Vacant lots

Run the function for all vacant lots with sufficient lot size to build.

```{r units-vacant}
units_vacant <- zoningdata1 %>% 
  filter(category == "vacant",
         smallsize_flag == 0) %>% 
  get_unit_variables(construction_efficiency = .7,
                     average_unit_size = 940)


zoningdata1 %>% 
  filter(category == "vacant",
         smallsize_flag == 0) %>% 
  get_unit_summary(construction_efficiency = .7,
                 average_unit_size = 940,
                 dwelling_type_breakdown = TRUE) %>% 
  knitr::kable()



sfa <- units_vacant %>% 
  filter(dwelling_type == "single family attached (ADU)")


sfd <- units_vacant %>% 
  filter(dwelling_type == "single family detached")

```



<!-- Get summary statistics for several different parameters -->

<!-- ```{r sum} -->

<!-- map_df(c(.7, .85), ~ get_unit_summary(dataset = filter(zoningdata1, -->
<!--                                                        category == "vacant", -->
<!--                                                        smallsize_flag == 0), -->
<!--                                       construction_efficiency = ., -->
<!--                                       average_unit_size = 940, -->
<!--                                       dwelling_type_breakdown = FALSE)) %>%  -->
<!--   knitr::kable() -->

<!-- ``` -->


<!-- Look at the distribution of units for lots with more than 1 unit, to check for extreme outliers. -->

<!-- ```{r scatter, echo = FALSE} -->

<!-- zoningdata1 %>%  -->
<!--   filter(category == "vacant", -->
<!--          smallsize_flag == 0) %>%  -->
<!--   get_unit_variables(construction_efficiency = .7, -->
<!--                      average_unit_size = 940) %>%  -->
<!--   filter(units_total > 1) %>%  -->
<!--   ggplot() + -->
<!--   geom_point(mapping = aes(lotsize_sf, units_total), -->
<!--              alpha = .5) + -->
<!--   scale_x_continuous(labels = scales::comma, -->
<!--                      expand = expand_scale(mult = c(0, 0.002)),  -->
<!--                      limits = c(0, 250000), -->
<!--                      breaks = 0:6 * 50000) + -->
<!--   scale_y_continuous(expand = expand_scale(mult = c(0, 0.002)),  -->
<!--                      limits = c(0, 45), -->
<!--                      breaks = 0:7 * 50) + -->
<!--   scatter_grid() + -->
<!--   labs(x = "Lot size (square feet)", y = "Total units") + -->
<!--   theme(plot.margin = margin(t = 20, r = 20)) -->

<!-- ``` -->


Look more closely at vacant lots < 50,000 SF

```{r scatter-small, echo = FALSE}

zoningdata1 %>% 
   filter(category == "vacant",
         smallsize_flag == 0) %>% 
  get_unit_variables(construction_efficiency = .7,
                     average_unit_size = 940) %>% 
  filter(lotsize_sf < 50000,
         units_total > 1) %>% 
  ggplot() +
  geom_point(mapping = aes(lotsize_sf, units_total,
                           color = dwelling_type),
             alpha = .8) +
  scale_x_continuous(labels = scales::comma,
                     expand = expand_scale(mult = c(0, 0.0002)), 
                     limits = c(0, 50000),
                     breaks = 0:5 * 10000) +
  scale_y_continuous(expand = expand_scale(mult = c(0, 0.0002)), 
                     limits = c(0, 60),
                     breaks = 0:4 * 15) +
  scatter_grid() +
  labs(x = "Lot size (square feet)", y = "Total units") +
  theme(plot.margin = margin(t = 20,
                             r = 20))



```

## Write out results

Write out the results to a csv. Include jurisdiction name in the title.

```{r write-results}

brsum <- zoningdata1 %>% 
  filter(category == "vacant",
         smallsize_flag == 0) %>% 
  get_unit_summary(construction_efficiency = .7,
                   average_unit_size = 940,
                   dwelling_type_breakdown = TRUE) 

# brsum %>% 
#   write_csv(paste0("L:/Libraries/RegHsg/Prog/Tables/",
#                    filepath,
#                    "-vacant-lots-analysis.csv"))
# 

```


```{r}
mf.br <- zoningdata %>% 
  filter(!is.na(dwelling_type),
         dwelling_type == "multiple family dwellings (ADU)",
         category == "vacant") %>% 
  group_by(propaddress) %>%
  count() %>% 
  nrow()


br.summary <- 
  c(nrow(jurfinal), 
    nrow(filter(jurfinal, category == "vacant")), 
    mf.br, 
    brsum %>% filter(dwelling_type == "multiple family dwellings (ADU)") %>% pull(lots) %>% sum(),
    brsum %>% filter(dwelling_type == "multiple family dwellings (ADU)") %>% pull(units_total) %>% sum())


br.breakdown <- brsum %>% 
  ungroup() %>% 
  mutate(dwelling_type = ifelse(dwelling_type %in% c("Townhouse", "Semi-detached"),
                                "sf attached",
                                dwelling_type)) %>% 
  group_by(dwelling_type) %>% 
  summarize_at(vars(lots, units_total) ,sum) %>% 
  filter(dwelling_type != "All") %>% 
  arrange(desc(units_total))



br.summary


br.breakdown %>% knitr::kable()

```


Clear global environment.

```{r}

rm(list=ls())

```
