==> Prog/impact-estimates/04_vacant-lots-montgomery.Rmd <==
---
title: "Regional Housing Framework"
subtitle: "Vacant lots analysis for Montgomery County"
author: ""
output:
  html_document:
    number_sections: FALSE
    self_contained: TRUE
    code_folding: show
    toc: TRUE
    toc_float: TRUE
    css: ../../www/web_report.css
    editor_options:
      chunk_output_type: console
---

<link rel="stylesheet" href="//fonts.googleapis.com/css?family=Lato" />

Library: RegHsg

Project: Regional Housing Framework

Author: Patrick Spauster
 
Version: R 3.5.1, RStudio 1.1.423

Last updated `r format(Sys.time(), '%B %d, %Y')`

Environment: Local Windows session (desktop)


```{r rmarkdown-setup, echo = FALSE}
knitr::opts_chunk$set(warning = FALSE)
knitr::opts_chunk$set(message = FALSE)
# knitr::opts_chunk$set(eval = FALSE)
```

## Description

The goal of this analysis is to determine how many units could be added by filling vacant lots with the most-dense developments allowed under the current zoning codes.
The two steps are:

1. Merge cleaned vacant lots data with zoning code data.
2. Calculate how many units are allowed one each vacant lot.

## Set-up
Load libraries and functions
```{r setup}
library(tidyverse)
library(sf)

```

Create directory for data exports
```{r dir}
if (!dir.exists("../../Data")) {
  dir.create("../../Data")
}

if (!dir.exists("L:/Libraries/RegHsg/Prog/Tables")) {
  dir.create("L:/Libraries/RegHsg/Prog/Tables")
}
```

Set FIPS code and filepath name
```{r fips}

currentfips <- "24031"
filepath <- "montgomery"
jdir <- paste0("L:/Libraries/RegHsg/Data/", filepath, "/")

source("../../Macros/read-montgomery.R")


```

Load in cleaned Black Knight data for `r str_to_title(filepath)` County. First is R dataset

```{r readR}
if (!exists("jur")) {
  jur <- readRDS(paste0(jdir, 
               "postcleaned-",
               filepath,
               "-data.Rdata"))
} else {
  warning(filepath, " data already read in")
}


count(jur, rockflag != 1, gaitflag != 1, othernotmoco !=1)

```


## Clean zoning code information

### Move file 

First- **manually** move file from Box to the `zoning-codes` directory in the `Doc` folder on the L drive. Name file accordingly:
"jursidiction-zoning-code.csv"

### Read file

Read in zoning code, rename variables so that they are machine readable

```{r file-check}

if (!file.exists(paste0("L:/Libraries/RegHsg/Doc/zoning-codes/",
                              filepath,
                              "-zoning-code.csv"))) {
  
  stop("Move zoning file for jurisdiction to L:/Libraries/RegHsg/Doc/zoning-codes directory")
}

```

```{r read-zoning-csv}
z <- cols(
  Jur = col_character(),
  Code = col_character(),
  `Zone` = col_character(),
  `Type of Dwelling` = col_character(),
  `Site Area` = col_number(),
  `Site area per dwelling unit (sq. ft.) (minimum)` = col_number(),
  `Lot area` = col_number(),
  `Lot area per dwelling unit (sq. ft.) (minimum)` = col_number(),
  `Lot Width (average minimum, ft)` = col_character(),
  `Lot Width building line (average minimum, ft)` = col_character(),
  `Height (ft)` = col_double(),
  `Site coverage maximum` = col_number(),
  `Lot coverage maximum` = col_number(),
  `Floor area minimum` = col_double(),
  `Maximum Density of Floor-Area-Ratio (FAR)` = col_number(),
  `Setbacks (front)` = col_character(),
  `Side (ft)` = col_character(),
  `Rear (ft)` = col_character(),
  `Density (units/acre)` = col_number(),
  `Massing Units/Row` = col_number(),
  `Open Space minimum percentage` = col_number(),
  Comments = col_character()
)

zoningcode <- read_csv(paste0("L:/Libraries/RegHsg/Doc/zoning-codes/",
                              filepath,
                              "-zoning-code.csv"),
                       col_types = z)

names(zoningcode) <- 
  c("juristdiction", "zoning_code", "zone", "dwelling_type",
    "site_area_min", "sitearea_per_unit", "lot_area_min", "lotarea_per_unit",
    "lot_width","lot_width_building", "height_ft",
    "site_coverage_max", "lot_coverage_max", "floor_area_min", "far_max",
    "setbacks_front", "setbacks_side",
    "setbacks_rear", "density", "massing", "open_space", "comments")

# fix variables read in as whole numbers (should be percent)
zoningcode <- zoningcode %>% 
  mutate_at(vars(lot_coverage_max, site_coverage_max, open_space),
            ~ . / 100)
rm(z)

```

### Creating dwelling_type hierarchy

For the vacant lots analysis, we want to determine the `dwelling_type` that allows for the most dense construction for that zone. 


#### Split jurisdictions

Create zoning for each jurisdiction so we can split the analysis

```{r zoning by juristdiction}

zoning_montgomery <- zoningcode %>% 
  filter(juristdiction == "Montgomery")


zoning_rockville <- zoningcode %>% 
  filter(juristdiction == "Rockville")

write_csv(zoning_rockville, 
          "L:/Libraries/RegHsg/Doc/zoning-codes/rockville-zoning-code-clean.csv")

zoning_gaithersburg <- zoningcode %>% 
  filter(juristdiction == "Gaithersburg")

write_csv(zoning_gaithersburg,
          "L:/Libraries/RegHsg/Doc/zoning-codes/gaithersburg-zoning-code-clean.csv")


```




*For Montgomery*

```{r list-codesM}

zoning_montgomery %>% 
  count(dwelling_type) %>% 
  knitr::kable()

```


Since the variables are not always comparable, the following order, from most to least dense, applies:

1. apartment
2. townhouse
3. duplex-over
4. duplex-side
5. detached house

multi-use and general have the same zoning specs as apartment. Change all of the multi-use and general zoning codes to apartments for simplicity.

```{r edit mont codes}

zoning_montgomery <- zoning_montgomery %>% 
  mutate(dwelling_type = ifelse(dwelling_type == "general"| dwelling_type == "multi-use", 
                                "apartment", 
                                dwelling_type))

```

Check to see if they were replaced

```{r check codesM}

zoning_montgomery %>% 
  count(dwelling_type) %>% 
  knitr::kable()

```


### Create factor variable
To use this hierarchy, we transform the `dwelling_type` variable into a factor, assign the factors to the appropriate level, (1-6, as listed above).
We will also create a designation for dwelling types that are restricted to a certain amount of units (one unit for "One Family Dwelling" and one for two units for "Duplex"). 

**NOTE** This will be VERY jurisdiction specific.

**NOTE** If there are no one or two unit limitations, still make the variable and set them all equal to zero. This allows the functionality of the calculation to be more uniform across counties.

```{r dwelling-factorM}

zoning_montgomery1 <- zoning_montgomery %>% 
  mutate(dwelling_type=ifelse(is.na(dwelling_type)==1, "all", dwelling_type)) %>% 
  mutate(dwelling_type = factor(dwelling_type,
                                levels = c("apartment",
                                           "townhouse",
                                           "duplex-over",
                                           "duplex-side",
                                           "detached house"
                                           )),
         one_unit_flag = ifelse(dwelling_type == "detached house",
                                1,
                                0),
         two_unit_flag = ifelse(dwelling_type == "duplex-over" | dwelling_type == "duplex-side",
                                1,
                                0))

```



*Cleaning and Calculation*

## Clean data for the whole juristdiction

Filter out really large lots that should not be part of the calculation. These properties create a huge number of properties, and upon looking at them on google Maps, do not appear buildable. Chris also did this filter in fairfax.

Fliter out rockville and gaithersburg

Count how many vacant lots

```{r filtering}

jur %>% 
  filter(lotsize_sf>1000000 & category=="vacant") %>% 
count(lotsize_sf, propaddress)

mont <- jur %>% 
  filter(rockflag != 1, gaitflag!=1, othernotmoco != 1) %>% 
  filter(lotsize_sf < 1000000) #filter out unreasonably large lots that mess up the calculation


count(mont, rockflag, gaitflag, othernotmoco)

count(mont, category == "vacant")

```

See how many lots have missing key variables

```{r missing}

mont %>% 
  mutate(num = NA) %>% 
  group_by(category_detail) %>% 
  summarize_at(vars(propaddress, lotsize_sf, zoning, num),
               ~ sum(is.na(.))) %>% 
  mutate_at(vars(propaddress, lotsize_sf, zoning),
            .funs = funs(percent = scales::percent(. / num))) %>% 
  knitr::kable(caption = "Number and percent of missing values")

```
## Merge with cleaned data

Now that we have selected the correct records from the zoning code file, we can merge with the cleaned property records data. First, we have to do a bit of cleaning of these codes, including getting rid of instances where there is a tie, and correcting some of the notation. We will create a new variable in order to archive the original `zoning` variable.

```{r bk-clean}


montfinal <- mont %>% 
  mutate(zoning_code = case_when(
    zoning %in% c("CR1.0","CR0.5", "CR2.0", "CR5.0", 
                  "cR3", "CR5", "CR3.0", "CR4.0") ~ "CR",
    # if there is no number after RT we go with the most restrictive code
    zoning %in% c("RT","RT6", "RT6.0") ~ "RT-6.0",
    zoning %in% c("RT8", "RT8.0") ~ "RT-8.0",
    zoning %in% c("RT10", "RT10.0") ~ "RT-10.0",
    zoning %in% c("RT125", "RT12", "RT12.5") ~ "RT-12.5",
    zoning %in% c("RT15", "RT15.0") ~ "RT-15.0",
    zoning %in% c( "CR2.0", "CR5.0", "CR3") ~ "CR",
    zoning %in% c("CRN0.5","CRN1.5", "CRN025", "CRN075") ~ "CRN",
    zoning %in% c("CRT0.5","CRT075", "CRT1.5", "CRT1.0", 
                  "CRT125", "CRT1", "CRT175", "CRT2.25", 
                  "CRT2.5", "CRT 1") ~ "CRT",
    zoning %in% c("R3", "R30") ~ "R-30",
    zoning %in% c("R200T", "R200") ~ "R-200",
    zoning %in% c("R20") ~ "R-20",
    zoning %in% c("R40") ~ "R-40",
    zoning %in% c("R90","R90TD") ~ "R-90",
    zoning %in% c("R60") ~ "R-60",
    zoning %in% c("R10") ~ "R-10",
    zoning %in% c("RE2") ~ "Re-2",
    zoning %in% c("RE2C") ~ "Re-2C",
    zoning %in% c("RE1") ~ "Re-1",
    zoning %in% c("RURAL", "RUR") ~ "R",
    TRUE ~ zoning
  ))

z <- count(montfinal, zoning_code)
           
```

Merge on zoning code variable. Test that all variables merged.

```{r zoning-merge}

zoningdata <- left_join(montfinal, zoning_montgomery1, by = "zoning_code")


count(zoningdata, category)
# count(zoningdata, zoning_code)
# count(zoningdata, lot_area_min)

```



```{r antijoin}

anti_join(montfinal, zoning_montgomery1, by = c("zoning_code")) %>% 
  count(zoning) %>% 
  knitr::kable()

```
All of these codes that didn't join are either non-residential, or they correspond to codes in Rockville and Gaithersburg, which are being calculated separately


## Fill in missing codes

1. fill in lot area minimums from site minimums
2. fill in lot coverage maximums from site coverage maximums
3. fill in per unit lot size when there is not general minimum lot size restriction
4. We also need to fill in missing lot coverage

5. and make two assumptions for townhouses that do not have lot coverage restrictions in the montgomery zoning code.
  1. The highest lot coverage for townhouses elsewhere in the codes is 40, so I assume a lot max of 40.
  2. The highest density allowed for townhouses elsewhere in the codes is 15.02 units/acre, so I assume this density.
  
6. finally, we need to fill in the lotsize_acre variable for those lot sizes that we filled in from the parcel file.

```{r fillin-for-caluculation}

zoningdata1 <- zoningdata %>% 
  mutate(lot_area_min = case_when(
    is.na(lot_area_min) ~ site_area_min,
    TRUE ~ lot_area_min)) %>% 
  mutate(lot_coverage_max = case_when(
    is.na(lot_coverage_max) ~ site_coverage_max,
    TRUE ~ lot_coverage_max)) %>% 
  mutate(lotarea_per_unit = case_when(
    is.na(lotarea_per_unit) ~ sitearea_per_unit,
    TRUE ~ lotarea_per_unit)) %>% 
  # Fill in lot coverage with site coverage and assume a maximum townhouse density of 40%
  mutate(lot_coverage_max = if_else(is.na(lot_coverage_max) & 
                                      dwelling_type == "townhouse" & 
                                      zoning_code %in% c("CRN", "CRT", "CR"), 
                                    .4, 
                                    lot_coverage_max )) %>% 
  # Assume a maximum townhouse density of 15.02
  mutate(density = if_else(is.na(density) &
                             dwelling_type == "townhouse" & 
                             zoning_code %in% c("CRN", "CRT", "CR"), 
                           15.02, 
                           density)) %>% 
  #i've filled in and modified a bunch of lotsize_sf, I make sure acres reflects that.
  mutate(lotsize_acres = lotsize_sf / 43560) 

# check that we got all the acres.
count(zoningdata1, is.na(lotsize_acres), is.na(lotsize_sf))

count(zoningdata1, category)

```



Filter out lots that are too small to build on and lots with missing zoning information, and select most dense zoning code for those that are big enough.

```{r slice-and-dice}


zoningdata2 <- zoningdata1 %>% 
  filter(!is.na(zoning_code)) %>% 
  mutate(smallsize_flag= ifelse(lotsize_sf >= lot_area_min, 
                                0, 
                                1)) %>% 
  mutate(smallsize_flag= ifelse(is.na(smallsize_flag), 
                                0, 
                                smallsize_flag)) %>% 
  group_by(propaddress, smallsize_flag) %>% 
  slice(which.min(dwelling_type)) %>% 
  ungroup() %>% 
  group_by(propaddress) %>% 
  slice(which.min(smallsize_flag)) %>% 
  ungroup()

# see how many observations should drop between steps
count(zoningdata1, is.na(zoning_code))

count(zoningdata2, smallsize_flag)

count(zoningdata2, category)

```


## Calculate Missing Units

There are several variables we will use to calculate how many units we can put on each vacant lot.

1. Buildable area: this is `lotsize_sf` x the coverage ratio (`lot_coverage_max`)
2. Living area: this is the buildable area x the construction efficiency. The lower bound estimate is 70%, the upper bound estimate is 85%.
3. Units per floor: this is the living area divided by the average unit size. We use 940 sq. feet, based on the American Housing Survey 2017 data for the DC metro area.
4. Units allowed based on height limits: This is the units per floor x the number of stories.
5. Units allowed based on lot size: this is the lot area divided by `lotarea_per_unit`
6. massing - the number of units allowed in one row of townhouses
7. FAR and FAR per unit - I calculate the number of units per floor and multiply it by the number of floors allowed by height
8. density - the number of units allowed per acre


We can create a function that will allow us to easily adjust these parameters and compare results.

```{r get-est}

#' Get maximum unit variables
#' @description this function generates the variables necessary to calculate the maximum number of units that can occupy one lot, per the by-right zoning code.
#'
#' @param dataset dataset with lot size and zoning variables 
#' @param construction_efficiency Assumption for construction efficiency
#' @param average_unit_size Assuption for average unit size- should be derived from the American Housing Survey. Given in square feet.
#'
#' @return returns a dataframe with new variables: `parameters`, `buildable_area`, `living_area`, `units_per_floor`, `floors`,`FAR_stories`,`stories`, `units_FAR`,  `units_height`, `units_lotsize`, `units_density`, and `units_total`
#'
#' @examples
get_unit_variables <- function(dataset, 
                               construction_efficiency, 
                               average_unit_size) {
  
  dataset %>% 
    mutate(parameters = paste0("CE: ", construction_efficiency * 100,
                               "% , Unit size: ", average_unit_size,
                               " sf"),
           lot_coverage_max = ifelse(is.na(lot_coverage_max), .75, lot_coverage_max),
           buildable_area_height = lotsize_sf * lot_coverage_max,
           living_area_height = buildable_area_height*construction_efficiency,
           units_per_floor_height = floor(living_area_height/average_unit_size),
           floors_height = floor(height_ft/10),
           units_height = units_per_floor_height * floors_height,
           units_lotsize = floor(lotsize_sf / lotarea_per_unit),
           perc_build_per_floor = ifelse(far_max <= lot_coverage_max, 
                                         far_max, 
                                         lot_coverage_max),
           buildable_area_far = perc_build_per_floor * lotsize_sf,
           living_area_far = buildable_area_far * construction_efficiency,
           units_per_floor_far = floor(living_area_far / average_unit_size),
           percent_per_floor_far = (units_per_floor_far*average_unit_size*1/construction_efficiency)/lotsize_sf,
           stories_far = ifelse(percent_per_floor_far!=0&!is.na(percent_per_floor_far),
                                floor(far_max/percent_per_floor_far),
                                0),
           subset = ifelse(is.na(lot_area_min), NA, (lotsize_sf/lot_area_min)),
           units_height = units_per_floor_height * floors_height,
           units_density = floor(density * lotsize_acres),
           units_lotsize = floor(lotsize_sf / lotarea_per_unit),
           units_far = (stories_far*units_per_floor_far) %>% as.numeric,
           units_far = ifelse(!is.na(far_max) & is.na(units_far), 0 , units_far)
          ) %>% mutate_at(vars(units_height, units_lotsize, units_far),
              funs(case_when(
                one_unit_flag == 1 ~ 1,
                two_unit_flag == 1 ~ 2,
                TRUE ~ .))) %>% 
    mutate(units_total = case_when(
       subset < 1 ~ 0,
       lotsize_sf < lot_area_min ~ 0,
       one_unit_flag == 1 ~ 1,
       two_unit_flag == 1 ~ 2,
       one_unit_flag == 1 & subset>=1 & units_lotsize/subset>=1 ~ 1*subset,
       two_unit_flag == 1 & subset>=1 & units_lotsize/subset>=2 ~ 2*subset,
       TRUE ~ pmin(units_lotsize, units_height, units_density, units_far, na.rm = TRUE)))
  
}

#' Get maximum unit summary
#'
#' @param dataset dataset with lot size and zoning variables 
#' @param construction_efficiency Assumption for construction efficiency
#' @param average_unit_size Assuption for average unit size- should be derived from the American Housing Survey. Given in square feet.
#' @param dwelling_type_breakdown Options are TRUE or FALSE- indicates whether or not estimates should be broken down by `dwelling_type`
#'
#' @return Returns a dataframe with summary statistics on the maximum units that can be build on lots, per the by-right zoning code.
#'
#' @examples
get_unit_summary <- function(dataset, 
                             construction_efficiency, 
                             average_unit_size,
                             dwelling_type_breakdown) {

  x <- dataset %>% 
    get_unit_variables(construction_efficiency,
                       average_unit_size) %>%
    filter(units_total > 0) %>% #i've added an additional filter to my function - there are several units where the lotsize passes the first filter, but the buildable area is not large enough to build on - so it creates 0 units. 
    group_by(parameters) %>% 
    summarize(lots = n(),
              units_height = sum(units_height, na.rm = TRUE),
              units_lotsize = sum(units_lotsize, na.rm = TRUE),
              units_far = sum(units_far, na.rm = TRUE),
              units_density = sum(units_density, na.rm = TRUE),
              units_total = sum(units_total, na.rm = TRUE),
              units_per_lot = floor(sum(units_total / lots, na.rm = TRUE)))
  
  y <- dataset %>% 
    get_unit_variables(construction_efficiency,
                       average_unit_size) %>%
    filter(units_total > 0) %>% 
    group_by(parameters, dwelling_type) %>% 
    summarize(lots = n(),
              units_height = sum(units_height, na.rm = TRUE),
              units_lotsize = sum(units_lotsize, na.rm = TRUE),
              units_far = sum(units_far, na.rm = TRUE),
              units_density = sum(units_density, na.rm = TRUE),
              units_total = sum(units_total, na.rm = TRUE),
              units_per_lot = floor(sum(units_total / lots, na.rm = TRUE))) %>% 
    mutate(dwelling_type = as.character(dwelling_type))
  
  
  if (dwelling_type_breakdown == TRUE) {
   
      bind_rows(y, mutate(x, dwelling_type = "All"))  
  
  } else if (dwelling_type_breakdown == FALSE) {
      x
    }
}



```


Create a dataset with results for troubleshooting purposes. Check to make sure there aren't more townhouses than are allowed to be built on any one property.


```{r sumassign}
resultsall <- get_unit_variables(dataset = zoningdata2,
                            construction_efficiency = .7,
                            average_unit_size = 940)

```


```{r testing 123}


testing123 <- resultsall %>% 
  filter(dwelling_type == "apartment"| dwelling_type == "townhouse", units_far == 0) %>% 
  select(units_total, dwelling_type, zoning_code, living_area_far, lot_coverage_max, far_max, stories_far , units_per_floor_far, units_far)
  #filter(zoning_code %in% c("CR", "CRT", "CRN"))

testing1234 <- resultsall %>% 
    filter(dwelling_type == "townshouse", zoning_code %in% c("CR", "CRT"), units_total == 0) %>% 
  select(units_total, dwelling_type, zoning_code, living_area_far, lot_coverage_max, far_max, stories_far , units_per_floor_far, units_far, units_lotsize, units_height, units_density, subset, lot_area_min )



```



Get summary statistics for several different parameters

```{r sumtable}

get_unit_summary(dataset = zoningdata2,
                 construction_efficiency = .7,
                 average_unit_size = 940,
                 dwelling_type_breakdown = TRUE) %>% 
  knitr::kable()

```


## Write out results

Write out the results to a csv. Include jurisdiction name in the title.

```{r write-results}

get_unit_summary(dataset = zoningdata2,#write out table
                 construction_efficiency = .7,
                 average_unit_size = 940,
                 dwelling_type_breakdown = TRUE) %>% 
  
  write_csv(paste0("L:/Libraries/RegHsg/Prog/Tables/",
                   filepath,
                   "-max-units.csv"))


get_unit_variables(dataset = zoningdata2, #write out file for sarah
                 construction_efficiency = .7,
                 average_unit_size = 940) %>% 
  
  saveRDS(paste0(jdir, 
               filepath,
               "max-units.Rdata"))


```

## Clean vacant lots data

Select only vacant lots, filter out for rockville and gaithersburg which will be calculated separately, and check that I got them all.

```{r resV}

vacant_mont <- zoningdata2 %>% 
  filter(category == "vacant", smallsize_flag == 0) #filter out all but vacant and small size flag

count(vacant_mont, rockflag, gaitflag) #make sure no rock or gaith
count(vacant_mont, category)
```


See how many lots have missing key variables

```{r missingV}

vacant_mont %>% 
  mutate(num = NA) %>% 
  group_by(category_detail) %>% 
  summarize_at(vars(propaddress, lotsize_sf, zoning, num),
               ~ sum(is.na(.))) %>% 
  mutate_at(vars(propaddress, lotsize_sf, zoning),
            .funs = funs(percent = scales::percent(. / num))) %>% 
  knitr::kable(caption = "Number and percent of missing values")

```

### Assessed Value
I fill in some missing values using the montgomer county shape file and then filter out any properties that have improvement values, implying there is a structure and not truly vacant.

```{r fill in assessed value data from the zoning file}

vacant_mont1 <- vacant_mont %>% 
  mutate(assessedlandvalue = if_else(is.na(assessedlandvalue), 
                                     LAND_ASSMT, 
                                     as.integer(assessedlandvalue))) %>% 
  mutate(assessedimprovementvalue = if_else(is.na(assessedimprovementvalue), 
                                            IMPROV_ASSMT, 
                                            as.integer(assessedimprovementvalue)))

sum(is.na(vacant_mont$assessedimprovementvalue)) - sum(is.na(vacant_mont$assessedimprovementvalue))


vacant_mont1 %>% 
  filter(assessedimprovementvalue > 0, 
         !is.na(assessedimprovementvalue)) %>% 
  mutate(p_total = assessedimprovementvalue / totalassessedvalue) %>% 
  group_by(category_detail) %>% 
  summarize(n = n(),
            mean = mean(assessedimprovementvalue),
            median = median(assessedimprovementvalue),
            min = min(assessedimprovementvalue),
            max = max(assessedimprovementvalue),
            p_total = median(p_total)) %>% 
  knitr::kable()

vacant_mont2 <- vacant_mont1 %>% 
  mutate(category = ifelse(assessedimprovementvalue > 0 & 
                           !is.na(assessedimprovementvalue),
                           "vacant with structure",
                           category))
count(vacant_mont2, category)
vacant_mont2 <- vacant_mont2 %>% 
  filter(category == "vacant")


```

Assign results to a dataset 
```{r sumVassign}
resultsvacant <- get_unit_variables(dataset = vacant_mont2,
                            construction_efficiency = .7,
                            average_unit_size = 940)

```
 

Get summary statistics for several different parameters

```{r sumV}

get_unit_summary(dataset = vacant_mont2,
                 construction_efficiency = .7,
                 average_unit_size = 940,
                 dwelling_type_breakdown = TRUE) %>% 
  knitr::kable()

brsum <- get_unit_summary(dataset = vacant_mont2,
                 construction_efficiency = .7,
                 average_unit_size = 940,
                 dwelling_type_breakdown = TRUE)


```




Look at the distribution of units for lots with more than 1 unit, to check for extreme outliers.

```{r scatter, echo = FALSE}

vacant_mont2 %>% 
  get_unit_variables(construction_efficiency = .7,
                     average_unit_size = 940) %>% 
  filter(units_total > 1) %>% 
  ggplot() +
  geom_point(mapping = aes(lotsize_sf, units_total),
             alpha = .5) +
  scale_x_continuous(labels = scales::comma,
                     expand = expand_scale(mult = c(0, 0.002)), 
                     limits = c(0, 300000),
                     breaks = 0:6 * 50000) +
  scale_y_continuous(expand = expand_scale(mult = c(0, 0.002)), 
                     limits = c(0, 100),
                     breaks = 0:7 * 50) +
  #scatter_grid() + # removed because of buggy urbnmpr
  labs(x = "Lot size (square feet)", y = "Total units") +
  theme(plot.margin = margin(t = 20, r = 20))

```


Look more closely at vacant lots < 50,000 SF

```{r scatter-small, echo = FALSE}

vacant_mont2 %>% 
  get_unit_variables(construction_efficiency = .7,
                     average_unit_size = 940) %>% 
  filter(lotsize_sf < 50000,
         units_total > 1) %>% 
  ggplot() +
  geom_point(mapping = aes(lotsize_sf, units_total,
                           color = dwelling_type),
             alpha = .8) +
  scale_x_continuous(labels = scales::comma,
                     expand = expand_scale(mult = c(0, 0.0002)), 
                     limits = c(0, 50000),
                     breaks = 0:5 * 10000) +
  scale_y_continuous(expand = expand_scale(mult = c(0, 0.0002)), 
                     limits = c(0, 60),
                     breaks = 0:4 * 15) +
  #scatter_grid() +# removed because of buggy urbnmpr
  labs(x = "Lot size (square feet)", y = "Total units") +
  theme(plot.margin = margin(t = 20,
                             r = 20))



```

## Write out results

Write out the results to a csv. Include jurisdiction name in the title.

```{r write-resultsV}

get_unit_summary(dataset = vacant_mont2,
                 construction_efficiency = .7,
                 average_unit_size = 940,
                 dwelling_type_breakdown = TRUE) %>% 
  
write_csv(paste0("L:/Libraries/RegHsg/Prog/Tables/",
                   filepath,
                   "-vacant-lots-analysis.csv"))


```



## Calculate values needed to fill table
```{r tableV}
mf.br <- zoningdata1 %>% 
  filter(!is.na(dwelling_type),
         dwelling_type == "apartment",
         category == "vacant") %>% 
  group_by(propaddress) %>%
  count() %>% 
  nrow()

br.summary <- 
  c(nrow(jur), 
    nrow(filter(jur, category == "vacant")), 
    mf.br, 
    brsum %>% filter(dwelling_type == "apartment") %>% pull(lots) %>% sum(),
    brsum %>% filter(dwelling_type == "apartment") %>% pull(units_total) %>% sum())


br.breakdown <- brsum %>% 
  ungroup() %>% 
  mutate(dwelling_type = ifelse(dwelling_type %in% c("townhouse", "duplex-over", "duplex-side"),
                                "sf attached",
                                dwelling_type)) %>% 
  group_by(dwelling_type) %>% 
  summarize_at(vars(lots, units_total),sum) %>% 
  filter(dwelling_type != "All") %>% 
  arrange(desc(units_total))


br.summary

br.breakdown %>% knitr::kable()

```

Clear global environment.

```{r}

rm(list=ls())

```
