==> Prog/impact-estimates/05_combine-parcels.Rmd <==
---
title: "Regional Housing Framework"
subtitle: "Combining Parcels"
author: ""
output:
  html_document:
    number_sections: FALSE
    self_contained: TRUE
    code_folding: show
    toc: TRUE
    toc_float: TRUE
    css: ../../www/web_report.css
    editor_options:
      chunk_output_type: console
---

<link rel="stylesheet" href="//fonts.googleapis.com/css?family=Lato" />

Library: RegHsg

Project: Regional Housing Framework

Author: Chris Davis, adapting from Sarah Strochak

Version: R 3.5.1, RStudio 1.1.423

Last updated `r format(Sys.time(), '%B %d, %Y')`

Environment: Local Windows session (desktop)


```{r rmarkdown-setup, echo = FALSE}
knitr::opts_chunk$set(warning = FALSE)
knitr::opts_chunk$set(message = FALSE)
# knitr::opts_chunk$set(eval = FALSE)
```
## Description

The goal of this analysis is to combine parcels for the upzoning analysis.

## Set-up

Load libraries and functions
```{r setup}
library(tidyverse)

source("../../Macros/read-fairfax.R")
source("../../Macros/read-arlington.R")
source("../../Macros/read-montgomery.R")
source("../../Macros/read-DC.R")

```

## Read each jurisdiction


Load in cleaned Black Knight data for all jurisdictions.

```{r readfair}
if (!exists("fairfax")) {
 fairfax <- read_fairfax()
} else {
  warning(filepath, " data already read in")
}

```


```{r read-dc}
if (!exists("dc_full")) {
 dc_full <- read_dc()
} else {
  warning(filepath, " data already read in")
}

```


```{r read-mon}
if (!exists("montgomery")) {
 montgomery <- readRDS("L:/Libraries/RegHsg/Data/montgomery/postcleaned-montgomery-data.Rdata")
} else {
  warning(filepath, " data already read in")
}

```


```{r read-arl}
if (!exists("arlington")) {
 arlington <- read_arlington()
} else {
  warning(filepath, " data already read in")
}

```


## Fill in variables

For each jurisdiction:

* Fill in missing variables
* Replace 0 with NA when appropriate
* Select variables needed for upzoning and soft sites analysis

### Missing units functions

Function to sum missing units 

```{r missing-fxn-units}

sum_missing_units <- function(dataset) {
  
  dataset %>% 
    group_by(category_detail) %>% 
    filter(residential == 1, category != "vacant") %>% 
    summarize(count = n(),
              units.na = sum(is.na(numberofunits)),
              units.zero = sum(numberofunits == 0, na.rm = TRUE),
              p.na = units.na / n(),
              p.zero = units.zero / n())
}


```

Function to sum missing lotsize 

```{r missing-fxn-ls}
sum_missing_lotsize <- function(dataset) {
  
  dataset %>% 
    group_by(category_detail) %>% 
    filter(residential == 1, category != "vacant") %>% 
    summarize(count = n(),
              lotsize.na = sum(is.na(lotsize_sf)),
              lotsize.zero = sum(lotsize_sf == 0, na.rm = TRUE),
              p.na = lotsize.na / n(),
              p.zero = lotsize.zero / n())
}


```

### Arlington

#### Number of units

Replace `0` with `NA` for multifamily

```{r arl-missing}

arlington %>% 
  sum_missing_units() %>% 
  knitr::kable()

arlington1 <- arlington %>% 
  mutate(numberofunits = ifelse(category == "mf" & numberofunits == 0,
                                NA,
                                numberofunits))

```



#### Lot size

Replace `0` with `NA` for all parcels.

```{r arl-ls}

arlington1 %>% 
  sum_missing_lotsize() %>% 
  knitr::kable()

arlington2 <- arlington1 %>% 
  mutate(lotsize_sf = ifelse(lotsize_sf == 0,
                             NA,
                             lotsize_sf))

```


### Fairfax

#### Number of units

```{r ff-tab}

fairfax %>% 
  sum_missing_units() %>% 
  knitr::kable()

```

Fill in duplex, SF, and townhouse.
```{r ff-sf}

fairfax1 <- fairfax %>% 
  mutate(numberofunits = ifelse(category_detail == "duplex" & numberofunits == 0,
                                2,
                                numberofunits),
         numberofunits = ifelse(category_detail == "townhouse" & address_type == "single",
                                1,
                                numberofunits),
         numberofunits = ifelse(category == "sf" & address_type == "single",
                                1,
                                numberofunits))

fairfax1 %>% 
  sum_missing_units() %>% 
  knitr::kable()

```

Look at all the additional units variables.

```{r ff-units}

fairfax %>% 
  filter(residential == 1, category != "vacant") %>% 
  group_by(category_detail) %>% 
  summarize_at(vars(numberofunits, num_livunit, dwellings_per_parcel,
                    units_area, numberofunits_tax),
               ~ sum(is.na(.)) / n()) %>% 
  knitr::kable()

```

Fill in with `num_livunit`, and then `numberofunits_tax` when possible.

```{r ff-mf}

fairfax2 <- fairfax1 %>% 
  mutate(numberofunits = ifelse(numberofunits == 0 & !is.na(num_livunit),
                                num_livunit,
                                numberofunits),
         numberofunits = ifelse(numberofunits == 0 & !is.na(numberofunits_tax),
                                numberofunits_tax,
                                numberofunits))

fairfax2 %>% 
  sum_missing_units() %>% 
  knitr::kable()

```

#### Lot size

```{r ff-ls}

fairfax2 %>% 
  sum_missing_lotsize() %>% 
  knitr::kable()

fairfax2 %>% 
  group_by(category_detail) %>% 
  summarize_at(vars(area_tax, parcel_area, sf_area),
               ~ sum(is.na(.)) / n()) %>% 
  knitr::kable()

```

Fill in where possible.

```{r ff-fill-ls}

fairfax3 <- fairfax2 %>% 
  mutate(lotsize_sf = ifelse(is.na(lotsize_sf) & !is.na(parcel_area) & parcel_area != 0,
                             parcel_area,
                             lotsize_sf),
         lotsize_sf = ifelse(is.na(lotsize_sf) & !is.na(area_tax) & area_tax != 0,
                             area_tax,
                             lotsize_sf),
         lotsize_sf = as.numeric(lotsize_sf),
         lotsize_acres = lotsize_sf / 43560)

fairfax3 %>% 
  sum_missing_lotsize() %>% 
  knitr::kable()

```

### Montgomery

#### Number of units

Fill in single family with 1

```{r mont-tab}

montgomery1 <- montgomery %>% 
  mutate(numberofunits = ifelse(category_detail == "sf detached" & 
                                  numberofunits == 0,
                                1,
                                numberofunits))

montgomery1 %>% 
  sum_missing_units() %>% 
  knitr::kable()

```

#### Lot size

```{r mont-ls}

montgomery1 %>% 
  sum_missing_lotsize() %>% 
  knitr::kable()

```

### DC

#### Number of units

Units were already filled for rental properties in `clean-DC.Rmd`. Fill in single family homes. Change `0` to `na` for apartments.

```{r dc-tab}

dc1 <- dc_full %>% 
  mutate(numberofunits = ifelse(is.na(numberofunits) & 
                                  category_detail == "Single-family home" &
                                  address_type %in% c("single", "missing"),
                                1,
                                numberofunits)) %>% 
  # align categories with other jurisdictions
  mutate(category = ifelse(category == "all other category",
                           "other",
                           category)) %>% 
  mutate(numberofunits = ifelse(category == "mf" & numberofunits == 0,
                                NA,
                                numberofunits))

dc1 %>% 
  sum_missing_units() %>% 
  knitr::kable()
  
```

#### Lot size

```{r dc-ls}

dc1 %>% 
  sum_missing_lotsize() %>% 
  knitr::kable()

```

## Select and combine

```{r}
arl <- arlington2 %>%
          select(county_fips,
                 county_name,
                 assessorsparcelnumberapnpin,
                 propaddress,
                 propcity,
                 address_type,
                 zoning, 
                 lotsize_sf, 
                 numberofunits,
                 category,
                 category_detail, 
                 countylandusedescription, 
                 residential, 
                 totalassessedvalue,
                 assessedimprovementvalue,
                 assessedlandvalue,
                 lat, 
                 long) %>% 
  mutate(numberofunits = as.integer(numberofunits))

fai <- fairfax3 %>%
        select(county_fips,
               county_name,
               assessorsparcelnumberapnpin,
               propaddress,
               propcity,
               address_type,
               zoning, 
               lotsize_sf, 
               numberofunits,
               category,
               category_detail, 
               countylandusedescription, 
               residential, 
               totalassessedvalue,
               assessedimprovementvalue,
               assessedlandvalue,
               lat, 
               long)

mon <- montgomery1 %>%
        select(county_fips,
               county_name,
               assessorsparcelnumberapnpin,
               propaddress,
               propcity,
               address_type,
               zoning, 
               lotsize_sf, 
               numberofunits,
               category,
               category_detail, 
               countylandusedescription, 
               residential, 
               totalassessedvalue,
               assessedimprovementvalue,
               assessedlandvalue,
               lat = latarc, 
               long = longarc)

dc <- dc1 %>%
    select(county_fips,
           county_name,
           assessorsparcelnumberapnpin = parcelbase_SSL,
           propcity,
           address_type,           
           zoning, 
           lotsize_sf, 
           numberofunits, 
           category,
           category_detail, 
           propaddress, 
           countylandusedescription, 
           residential,
           totalassessedvalue,
           assessedimprovementvalue,
           assessedlandvalue,
           lat, long,
           parcelgeo_x, parcelgeo_y,
           parcelbase_address)

```

Bind rows together
```{r}

all_jurs <- arl %>% 
  bind_rows(fai) %>%
  bind_rows(mon) %>%
  bind_rows(dc) %>% 
  # make sure there are no 0s, only NAs
  mutate_at(vars(lotsize_sf, numberofunits),
            ~ ifelse(. == 0, NA, .)) %>% 
  mutate(lotsize_acres = lotsize_sf / 43560)
            

```

Look at combined missing variables.

```{r com-missing}

all_jurs %>% 
  sum_missing_units() %>% 
  knitr::kable(caption = "Missing units for all jurisdictions")


all_jurs %>% 
  sum_missing_lotsize() %>% 
  knitr::kable(caption = "Missing lot size for all jurisdictions")


all_jurs %>% 
  filter(category != "vacant", residential == 1) %>% 
  summarize(sum(is.na(numberofunits)) / n(),
            sum(is.na(lotsize_sf))/ n())

```


Write out data
```{r write-all}

write_csv(all_jurs, "L:/Libraries/RegHsg/Data/parcel-all-cleaned-data.csv")

```

Clear global environment.

```{r}

rm(list=ls())

```
